dev-java/kotlin-common-bin: Expand and make use of MAVEN_PROVIDES
With this change, MAVEN_PROVIDES is coupled with the list of JARs
installed by the ebuild, so all JARs installed are listed in
MAVEN_PROVIDES, and all artifacts in MAVEN_PROVIDES are actually
installed.  java-ebuilder will be capable of using this ebuild as
dependency for all Kotlin core library components so no extraneous
ebuilds for Kotlin library artifacts will be generated.

Package-Manager: Portage-3.0.18, Repoman-3.0.2
Signed-off-by: Yuan Liao <liaoyuan@gmail.com>