dev-lang/kotlin-*: Generate MANIFEST.MF for runtime components
This is required to prevent the Kotlin compiler from complaining about
bundled Kotlin Runtime library.  The packages concerned by this commit
are literally the Runtime libraries themselves so the compiler will emit
such false warnings, but the upstream suppresses them by adding special
fields to MANIFEST.MF to indicate this, and if the compiler is able to
find those fields, it will not show the warning.

Signed-off-by: Yuan Liao <liaoyuan@gmail.com>