dev-java/hibernate-validator: Depend on openjfx only when USE="binary"
Although the jdeps tool says that this package's JAR depends on JavaFX,
JavaFX is never used in the classpath of spark-demo during the runtime,
and it might be just fine to declare it as a build-only dependency.
JavaFX has lots of transitive dependencies that can make the total merge
time very long, so it would be great if it does not get pulled
unnecessarily.

Package-Manager: Portage-3.0.20, Repoman-3.0.2
Signed-off-by: Yuan Liao <liaoyuan@gmail.com>