dev-java/maven-artifact-manager: Get rid of last-rited dependency
dev-java/backport-util-concurrent, a dependency of this package, will be
removed from ::gentoo soon.  The dependency is basically just a
relocated version of java.util.concurrent, which has been available in
JDK for very long time.  It should be possible to just switch to JDK's
java.util.concurrent instead of continue relying on the relocated
version.

Bug: https://bugs.gentoo.org/819507
Signed-off-by: Yuan Liao <liaoyuan@gmail.com>