.github/actions/run-test-case: Update ebuild-commander setup command
Following setuptools's './setup.py install' deprecation,
ebuild-commander is going to drop setup.py in its source tree in the
next release and rely on pip for project installation.  Although the
current latest release, v0.1.0, can still be installed via './setup.py
install', it can also be installed with pip.  Switching the installation
command is not necessary at this time, but the change will be required
in the future, so it does not hurt to make the move right now.

Signed-off-by: Yuan Liao <liaoyuan@gmail.com>