tests: Fix errors caused by libffi 3.4 being pulled as dependency
As dev-java/jna-5.10.0 has landed in ::gentoo, >=dev-libs/libffi-3.4,
one of its dependencies, is now in the dependency tree of packages in
this overlay.  Any package that depends on libffi is to be rebuilt upon
libffi's sub-slot change, but the rebuilds cannot be triggered by
merging jna-5.10.0.  Therefore, we manually trigger the rebuilds by
explicitly merging >=dev-libs/libffi-3.4 in a dedicated 'emerge'
invocation.

Signed-off-by: Yuan Liao <liaoyuan@gmail.com>