dev-java/duke: Move build-only dependencies out of CP_DEPEND
Duke is in this overlay solely to support H2O as a dependency.  Duke
depends on Servlet API 2.4 as a build-only dependency, whereas H2O
transitively has Servlet API 3.1 as a runtime dependency.  If Servlet
2.4 is declared as a classpath dependency (build + runtime) of Duke, it
will be added to H2O's runtime classpath, causing two incompatible
versions of Servlet API to be included in the classpath, which may lead
to strange runtime errors.  Therefore, it is important to eliminate
redundant dependencies from CP_DEPEND.

Signed-off-by: Yuan Liao <liaoyuan@gmail.com>