kotlin-utils.eclass: Make `KOTLIN_UTILS_SELECTED_VERSION` public
This variable is useful for ebuilds when there are tasks that should be
done conditionally on the feature release version of Kotlin compiler
used to build it, like applying patches, modifying compiler arguments,
etc.  Getting the version by checking which KOTLIN_SINGLE_TARGET
USE_EXPAND flag is enabled seems to be feasible, but that would no
longer indicate the correct version if KOTLIN_COMPAT_OVERRIDE is in
effect because KOTLIN_COMPAT_OVERRIDE does not alter USE settings.

Signed-off-by: Yuan Liao <liaoyuan@gmail.com>