virtual/kotlin: Reorder providers to avoid potential dep resol issue
As of Portage 3.0.30, the first package in an any-of group will be
selected if no packages in the group are installed.  Because of this,
putting the imaginary dev-lang/kotlin before dev-lang/kotlin-bin might
cause Portage to complain about circular dependency if it is created in
the future and depends on dev-lang/kotlin-bin.  virtual/jdk and
virtual/jre in ::gentoo also list a binary package before its
corresponding source-based package.

Signed-off-by: Yuan Liao <liaoyuan@gmail.com>