kotlin-{utils,libs}.eclass: Support limiting a feature rel to lib build
kotlin-utils.eclass, kotlin-libs.eclass: Support limiting a Kotlin
feature release to building Kotlin library packages only

Kotlin library packages that inherit kotlin-libs.eclass use
KOTLIN_COMPAT to declare the Kotlin compiler feature releases that can
be used to build them too, although in this case KOTLIN_COMPAT usually
just contains one element, which is the identifier for the feature
release the library packages belongs to.

Even if a certain Kotlin feature release is no longer eligible for
KOTLIN_SINGLE_TARGET, it might still be useful to keep the compiler and
library packages for it in the repository for uses outside Portage.
This would essentially require two separate sets of KOTLIN_COMPAT
processing logic, for Kotlin library packages and third-party Kotlin
packages respectively.  This allows a KOTLIN_SINGLE_TARGET to be used
exclusively for Kotlin library packages, so the compiler and library
packages for the feature release may still be installed, and at the same
time, third-party Kotlin packages cannot use it.

Signed-off-by: Yuan Liao <liaoyuan@gmail.com>