dev-java/netty: Move slot 0 -> 3.10
A version of this package that uses slot 0 and is incompatible with
3.10.6 has been added to ::gentoo, so the existing version in this
overlay has to move to a different slot.

A slot move for a Java package usually requires a revision bump because
for Java packages, installation paths of some files (such as the JARs)
are dependent upon the slot.

Signed-off-by: Yuan Liao <liaoyuan@gmail.com>